# docs/api/api_gateway/api_gateway_test/openapi.yaml
openapi: 3.0.0
info:
  contact:
    name: ""
  description: "The Articles API provides endpoints for managing and retrieving articles\
    \ and article versions within the wiki application. It supports core CRUD (Create,\
    \ Read, Update, Delete) operations, search functionality, and versioning."
  license:
    name: ""
  title: laWiki API Gateway
  version: "2.0"
servers:
- description: SwaggerHub API Auto Mocking
  url: https://virtserver.swaggerhub.com/JABR3830/ArticlesAPI/1.0
- description: SwaggerHub API Auto Mocking
  url: https://virtserver.swaggerhub.com/JABR3830/ArticleAPI/1.0
paths:
  /v2/articles/{id}:
    delete:
      description: Delete an Article identified by it's unique ID
      operationId: deleteArticleByIDV2
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          description: Article successfully deleted.
        "400":
          description: "Bad Request, invalid Article ID format. "
        "403":
          description: Forbidden
        "404":
          description: Article Not Found
      security:
      - oauth_token: []
      summary: Delete Article
      tags:
      - v2/editors
    get:
      description: Get an Article identified by it's unique ID
      operationId: getArticleByIdV2
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              examples:
                Example 1:
                  value:
                    id: 497f6eca-6276-4993-bfeb-53cbbbba6f08
                    wiki_id: 9a3179cd-aa83-4a5b-a08a-5c794d4020a8
                    title:
                      property1: string
                      property2: string
                    creation_date: 2019-08-24T14:15:22Z
                    author:
                      id: 497f6eca-6276-4993-bfeb-53cbbbba6f08
                      name: string
                      image: http://example.com
                    tags:
                    - id: 497f6eca-6276-4993-bfeb-53cbbbba6f08
                      tag: ac5f6b7f-bb01-4c1e-a5c0-7a9f62d81c22
                      wiki_id: 9a3179cd-aa83-4a5b-a08a-5c794d4020a8
                      articles:
                      - id: 497f6eca-6276-4993-bfeb-53cbbbba6f08
                        name: string
                      translations:
                        property1: string
                        property2: string
                    versions:
                    - id: 497f6eca-6276-4993-bfeb-53cbbbba6f08
                      title:
                        property1: string
                        property2: string
                      author:
                        id: 497f6eca-6276-4993-bfeb-53cbbbba6f08
                        name: string
                        image: http://example.com
                      modification_date: 2019-08-24T14:15:22Z
                      lan: string
                    rating: 0
                    lan: string
              schema:
                $ref: '#/components/schemas/Article'
          description: OK
        "400":
          description: "Bad Request, invalid Article ID format. "
        "404":
          description: Article Not Found
      summary: Get Article
      tags:
      - v2/public
...
  /v2/articles:
    get:
      description: "Get a list of Articles from a given Wiki that match a keyword\
        \ string. Results can by filtered by tags, sorted by different parameters\
        \ and support pagination."
      operationId: searchArticlesV2
      parameters:
      - description: The ID of the wiki where the serch will be made
        explode: true
        in: query
        name: wiki_id
        required: false
        schema:
          format: uuid
          type: string
        style: form
      - description: Search query for the name of the article
        explode: true
        in: query
        name: name
        required: false
        schema:
          example: name
          type: string
        style: form
      - description: A comma-separated list of tag IDs to search for
        explode: false
        in: query
        name: tags
        required: false
        schema:
          example:
          - tag1
          - tag2
          items:
            type: string
          type: array
          uniqueItems: true
        style: form
      - description: The index of the first result to return. Use with limit to get
          the next page of search results.
        explode: true
        in: query
        name: offset
        required: false
        schema:
          default: 0
          example: 10
          minimum: 0
          type: integer
        style: form
      - description: The maximum number of results to return.
        explode: true
        in: query
        name: limit
        required: false
        schema:
          default: 20
          example: 20
          maximum: 100
          minimum: 0
          type: integer
        style: form
      - description: Sorts the articles by different criteria
        explode: true
        in: query
        name: order
        required: false
        schema:
          default: none
          example: "recent, oldest, popular, unpopular"
          type: string
        style: form
      - description: Single date or range
        explode: true
        in: query
        name: creation_date
        required: false
        schema:
          format: date
          type: string
        style: form
      - description: Filter for the author of the Article
        explode: true
        in: query
        name: author_name
        required: false
        schema:
          type: string
        style: form
      - description: Filter for the editors of the Article
        explode: true
        in: query
        name: editor_name
        required: false
        schema:
          type: string
        style: form
      - description: "Language of the ArticleVersion, if none, the original language\
          \ will be returned"
        example: "en, es, fr"
        explode: true
        in: query
        name: lan
        required: false
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              examples:
                Example 1:
                  value:
                    articles:
                    - id: 497f6eca-6276-4993-bfeb-53cbbbba6f08
                      wiki_id: 9a3179cd-aa83-4a5b-a08a-5c794d4020a8
                      title:
                        property1: string
                        property2: string
                      creation_date: 2019-08-24T14:15:22Z
                      author:
                        id: 497f6eca-6276-4993-bfeb-53cbbbba6f08
                        name: string
                        image: http://example.com
                      tags:
                      - id: 497f6eca-6276-4993-bfeb-53cbbbba6f08
                        tag: ac5f6b7f-bb01-4c1e-a5c0-7a9f62d81c22
                        wiki_id: 9a3179cd-aa83-4a5b-a08a-5c794d4020a8
                        articles:
                        - id: 497f6eca-6276-4993-bfeb-53cbbbba6f08
                          name: string
                        translations:
                          property1: string
                          property2: string
                      versions:
                      - id: 497f6eca-6276-4993-bfeb-53cbbbba6f08
                        title:
                          property1: string
                          property2: string
                        author:
                          id: 497f6eca-6276-4993-bfeb-53cbbbba6f08
                          name: string
                          image: http://example.com
                        modification_date: 2019-08-24T14:15:22Z
                        lan: string
                      rating: 0
                      lan: string
                    total: 0
                    offset: 0
                    limit: 0
                    previous: http://example.com
                    next: http://example.com
              schema:
                $ref: '#/components/schemas/ArticleList'
          description: OK
        "400":
          content: {}
          description: "Bad Request, invalid input paramaters"
        "404":
          description: Article Not Found
      summary: Search for Articles
      tags:
      - v2/public
    post:
      description: Create a new Article in a given wiki
      operationId: createArticleV2
      requestBody:
        content:
          application/json:
            examples:
              Example 1:
                value:
                  title: string
                  author:
                    id: 497f6eca-6276-4993-bfeb-53cbbbba6f08
                    name: string
                    image: http://example.com
                  tags:
                  - id: 497f6eca-6276-4993-bfeb-53cbbbba6f08
                    tag: ac5f6b7f-bb01-4c1e-a5c0-7a9f62d81c22
                    wiki_id: 9a3179cd-aa83-4a5b-a08a-5c794d4020a8
                    articles:
                    - id: 497f6eca-6276-4993-bfeb-53cbbbba6f08
                      name: string
                    translations:
                      property1: string
                      property2: string
                  body: string
                  wiki_id: string
                  lan: string
                  trasnlate_title: true
            schema:
              $ref: '#/components/schemas/NewArticle'
      responses:
        "201":
          content:
            application/json:
              examples:
                Example 1:
                  value:
                    id: 497f6eca-6276-4993-bfeb-53cbbbba6f08
                    wiki_id: 9a3179cd-aa83-4a5b-a08a-5c794d4020a8
                    title:
                      property1: string
                      property2: string
                    creation_date: 2019-08-24T14:15:22Z
                    author:
                      id: 497f6eca-6276-4993-bfeb-53cbbbba6f08
                      name: string
                      image: http://example.com
                    tags:
                    - id: 497f6eca-6276-4993-bfeb-53cbbbba6f08
                      tag: ac5f6b7f-bb01-4c1e-a5c0-7a9f62d81c22
                      wiki_id: 9a3179cd-aa83-4a5b-a08a-5c794d4020a8
                      articles:
                      - id: 497f6eca-6276-4993-bfeb-53cbbbba6f08
                        name: string
                      translations:
                        property1: string
                        property2: string
                    versions:
                    - id: 497f6eca-6276-4993-bfeb-53cbbbba6f08
                      title:
                        property1: string
                        property2: string
                      author:
                        id: 497f6eca-6276-4993-bfeb-53cbbbba6f08
                        name: string
                        image: http://example.com
                      modification_date: 2019-08-24T14:15:22Z
                      lan: string
                    rating: 0
                    lan: string
              schema:
                $ref: '#/components/schemas/Article'
          description: Created
        "400":
          description: "Bad Request, invalid paramaters"
        "403":
          description: Forbidden
      security:
      - oauth_token: []
      summary: Create Article
      tags:
      - v2/editors
  /v2/articles/versions/{id}:
    delete:
      description: Delete an ArticleVersion identified by it's unique ID
      operationId: deleteArticleVersionByIdV2
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          description: ArticleVersion successfully deleted.
        "400":
          description: "Bad Request, invalid ArticleVersion ID format. "
        "403":
          description: Forbidden
        "404":
          description: ArticleVersion Not Found
      security:
      - oauth_token: []
      summary: Delete ArticleVersion
      tags:
      - v2/editors
    get:
      description: Get an ArticleVersion identified by it's unique ID
      operationId: getArticleVersionByIdV2
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      - description: "Language of the ArticleVersion, if none, the original language\
          \ will be returned"
        example: "en, es, fr"
        explode: true
        in: query
        name: lan
        required: false
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              examples:
                Example 1:
                  value:
                    id: 497f6eca-6276-4993-bfeb-53cbbbba6f08
                    article_id: e4aad634-3311-4e79-b165-3911549b3b29
                    author:
                      id: 497f6eca-6276-4993-bfeb-53cbbbba6f08
                      name: string
                      image: http://example.com
                    title:
                      property1: string
                      property2: string
                    modification_date: 2019-08-24T14:15:22Z
                    body: string
                    tags:
                    - id: 497f6eca-6276-4993-bfeb-53cbbbba6f08
                      tag: ac5f6b7f-bb01-4c1e-a5c0-7a9f62d81c22
                      wiki_id: 9a3179cd-aa83-4a5b-a08a-5c794d4020a8
                      articles:
                      - id: 497f6eca-6276-4993-bfeb-53cbbbba6f08
                        name: string
                      translations:
                        property1: string
                        property2: string
                    lan: string
              schema:
                $ref: '#/components/schemas/ArticleVersion'
          description: OK
        "400":
          description: "Bad Request, invalid ArticleVersion ID format. "
        "404":
          description: Article Version Not Found
      summary: Get ArticleVersion
      tags:
      - v2/public
  /v2/articles/{id}/versions:
    get:
      description: Get a list of ArticleVersions of a given Article. Results can be
        sorted by creation date adn support pagination.
      operationId: getArticleVersionListByArticleIDV2
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      - description: The index of the first result to return. Use with limit to get
          the next page of search results.
        explode: true
        in: query
        name: offset
        required: false
        schema:
          default: 0
          example: 10
          type: integer
        style: form
      - description: The maximum number of results to return.
        explode: true
        in: query
        name: limit
        required: false
        schema:
          default: 20
          example: 50
          maximum: 100
          minimum: 0
          type: integer
        style: form
      - description: Sorts the articles by different criteria
        explode: true
        in: query
        name: order
        required: false
        schema:
          example: "recent, oldest"
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              examples:
                Example 1:
                  value:
                    article_versions:
                    - id: 497f6eca-6276-4993-bfeb-53cbbbba6f08
                      title:
                        property1: string
                        property2: string
                      author:
                        id: 497f6eca-6276-4993-bfeb-53cbbbba6f08
                        name: string
                        image: http://example.com
                      modification_date: 2019-08-24T14:15:22Z
                      lan: string
                    total: 0
                    offset: 0
                    limit: 0
                    previous: http://example.com
                    next: http://example.com
              schema:
                $ref: '#/components/schemas/ArticleVersionList'
          description: OK
        "400":
          description: "Bad Request, invalid input paramaters"
        "404":
          description: Article not Found
      summary: Get Article's ArticleVersions
      tags:
      - v2/public
    post:
      description: Create an ArticleVersion for a given Article and adds it to the
        list of versions of the Article.
      operationId: createArticleVersionV2
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            examples:
              Example 1:
                value:
                  article_id: e4aad634-3311-4e79-b165-3911549b3b29
                  title: string
                  author:
                    id: 497f6eca-6276-4993-bfeb-53cbbbba6f08
                    name: string
                  tags:
                  - id: 497f6eca-6276-4993-bfeb-53cbbbba6f08
                    tag: string
                  body: string
            schema:
              $ref: '#/components/schemas/NewArticleVersion'
      responses:
        "201":
          content:
            application/json:
              examples:
                Example 1:
                  value:
                    id: 497f6eca-6276-4993-bfeb-53cbbbba6f08
                    article_id: e4aad634-3311-4e79-b165-3911549b3b29
                    author:
                      id: 497f6eca-6276-4993-bfeb-53cbbbba6f08
                      name: string
                      image: http://example.com
                    title:
                      property1: string
                      property2: string
                    modification_date: 2019-08-24T14:15:22Z
                    body: string
                    tags:
                    - id: 497f6eca-6276-4993-bfeb-53cbbbba6f08
                      tag: ac5f6b7f-bb01-4c1e-a5c0-7a9f62d81c22
                      wiki_id: 9a3179cd-aa83-4a5b-a08a-5c794d4020a8
                      articles:
                      - id: 497f6eca-6276-4993-bfeb-53cbbbba6f08
                        name: string
                      translations:
                        property1: string
                        property2: string
                    lan: string
              schema:
                $ref: '#/components/schemas/ArticleVersion'
          description: Version successfully created.
        "400":
          description: "Bad Request, Invalid Parameters"
        "403":
          description: Forbidden
      security:
      - oauth_token: []
      summary: Create ArticleVersion for an Article
      tags:
      - v2/editors
  /v2/articles/{article_id}/versions/{version_id}:
    put:
      description: Restore an older ArticleVersion of an Article.
      operationId: restoreArticleVersionV2
      parameters:
      - explode: false
        in: path
        name: article_id
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: version_id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          description: ArticleVersion successfully restored
        "400":
          description: invalid Article or ArticleVersion ID format
        "403":
          description: Forbidden
        "404":
          description: Article Version Not Found
      security:
      - oauth_token: []
      summary: Restore ArticleVersion
      tags:
      - v2/editors
...
  /v2/tags/articles/{id}:
    delete:
      description: "Unassigns a list of tags, given their IDs to an article."
      operationId: unassign-tagsV2
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      - description: List of Tag IDs
        explode: false
        in: query
        name: ids
        required: true
        schema:
          example:
          - tag1
          - tag2
          items:
            type: string
          maxItems: 50
          type: array
          uniqueItems: true
        style: form
      responses:
        "204":
          description: "No Content, tag unassigned succesfully"
        "400":
          description: "Bad Request, invalid invalid paramaters format"
        "403":
          description: Forbidden
        "404":
          description: Article or tag not found
      security:
      - oauth_token: []
      summary: Unassign Tags
      tags:
      - v2/editors
    get:
      description: Retrieves all the tags from an article.
      operationId: get-articles-tagsV2
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      - description: Maximum amount of responses to be returned
        explode: true
        in: query
        name: limit
        required: false
        schema:
          default: 20
          example: 20
          maximum: 100
          minimum: 0
          type: integer
        style: form
      - description: The index of the first result to return. Use with limit to get
          the next page of search results.
        explode: true
        in: query
        name: offset
        required: false
        schema:
          default: 0
          example: 0
          minimum: 0
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              examples:
                Example 1:
                  value:
                    articles:
                    - id: 497f6eca-6276-4993-bfeb-53cbbbba6f08
                      tag: ac5f6b7f-bb01-4c1e-a5c0-7a9f62d81c22
                      wiki_id: 9a3179cd-aa83-4a5b-a08a-5c794d4020a8
                      articles:
                      - id: 497f6eca-6276-4993-bfeb-53cbbbba6f08
                        name: string
                      translations:
                        property1: string
                        property2: string
                    total: 0
                    offset: 0
                    limit: 0
                    previous: http://example.com
                    next: http://example.com
              schema:
                $ref: '#/components/schemas/TagList'
          description: OK
        "404":
          description: Article not found
      summary: Get Articles Tag
      tags:
      - v2/public
    put:
      description: "Assigns a list of tags, given their IDs, to an article."
      operationId: assign-tagsV2
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            examples:
              Example 1:
                value:
                  tag_ids:
                  - 497f6eca-6276-4993-bfeb-53cbbbba6f08
            schema:
              $ref: '#/components/schemas/TagIDList'
      responses:
        "204":
          description: "No Content, tags assigned"
        "400":
          description: "Bad Request, invalid paramaters format"
        "403":
          description: Forbidden
        "404":
          description: Article or tag not found
      summary: Assign Tags
      tags:
      - v2/editors
...
  /v2/upload-image:
    post:
      description: Uploads an image file and returns the URL.
      operationId: uploadImageV2
      requestBody:
        content:
          multipart/form-data:
            examples:
              Example 1:
                value:
                  file: string
            schema:
              $ref: '#/components/schemas/v1_uploadimage_body'
      responses:
        "200":
          content:
            application/json:
              examples:
                Example 1:
                  value:
                    url: string
              schema:
                $ref: '#/components/schemas/inline_response_200'
          description: Successfully uploaded
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
      security:
      - oauth_token: []
      summary: Upload Image
      tags:
      - v2/editors
...
  /v1/articles/{article_id}/versions/{version_id}:
    put:
      description: Restore an older ArticleVersion of an Article.
      operationId: restoreArticleVersion
      parameters:
      - explode: false
        in: path
        name: article_id
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: version_id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          description: ArticleVersion successfully restored
        "400":
          description: invalid Article or ArticleVersion ID format
        "403":
          description: Forbidden
        "404":
          description: Article Version Not Found
      summary: Restore ArticleVersion
      tags:
      - v1/editors
...
  /v1/upload-image:
    post:
      description: Uploads an image file and returns the URL.
      operationId: uploadImage
      requestBody:
        content:
          multipart/form-data:
            examples:
              Example 1:
                value:
                  file: string
            schema:
              $ref: '#/components/schemas/v1_uploadimage_body'
      responses:
        "200":
          content:
            application/json:
              examples:
                Example 1:
                  value:
                    url: string
              schema:
                $ref: '#/components/schemas/inline_response_200'
          description: Successfully uploaded
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
      summary: Upload Image
      tags:
      - v1/editors
components:
...
  schemas:
...
    TagIDList:
      example:
        tag_ids:
        - 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        - 046b6c7f-0b8a-43b9-b35d-6489e6daee91
      properties:
        tag_ids:
          description: List of Tag IDs
          items:
            format: uuid
            type: string
          title: tag_ids
          type: array
          x-stoplight:
            id: kmi9kpdl1ztoh
      required:
      - tag_ids
      title: TagIDsList
      type: object
      x-examples:
        Example 2:
          tag_ids:
          - 497f6eca-6276-4993-bfeb-53cbbbba6f08
...
    v1_uploadimage_body:
      properties:
        file:
          format: binary
          type: string
      type: object
    inline_response_200:
      example:
        url: url
      properties:
        url:
          description: URL of the uploaded image
          title: url
          type: string
      title: inline_response_200
      type: object
  securitySchemes:
    oauth_token:
      scheme: bearer
      type: http